Made changes to allow Solaris5 platform to compile with gcc-warnings enabled
git-svn-id: svn://clusterresources.com/torque/branches/2.4-fixes@3393 8f940c70-5916-0410-ac46-d1fa2fa6ea29

+++ src/resmom/prolog.c
@@ -413,7 +413,7 @@ int run_pelog(
           (pelog != NULL) ? pelog : "NULL",
           (pjob != NULL) ? pjob->ji_qs.ji_jobid : "NULL",
           getcwd(tmpBuf, sizeof(tmpBuf)),
-          getpid());
+          (int)getpid());
 
         log_record(PBSEVENT_SYSTEM, 0, id, log_buffer);
         }
